extracted pools module
